backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# LDAP
python-ldap==3.4.4

# Redis
redis==5.0.1
aioredis==2.0.1

# Scheduling
apscheduler==3.10.4

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
aiofiles==23.2.1

# Compression
gzip-stream==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
